Add basic database instructions.
Add initial 'marketdraft' database creation instructions for developers.  This will not be necessary once developing in a cloud environment, but since local development is our starting place, we need to make sure the code sees a consistent view of the database.  The initial test data should be loaded by migrations (coming soon!).