Create users table and signup skeleton.
Create the structure of the users table, added to a migrations directory to
enable easy changes and rollbacks as necessary.  Also create the corresponding
db model in our application server code and add placeholders for the actual
login/signup pages.

Leave the actual form generation, parsing and validation for a future CL.