Add template functionalit to html pages.
Mostly copied from "Let's Go". Add a templateData struct (empty to start) to
provide dynamic data to the html rendering for pages.  Also add a template cache
to prevent the need to parse UI templates with every request. Finally, add a
render method to reduce code duplication for page rendering in preparation for
adding another page to the website (in addition to the home page).