Add basic user account handling: sign-up, login and logout.
Add forms, validation and db handling for the basic user model present in
"Let's Go".  This enables the application to be developed with the notion of a
user and to show user-specific information when logged in (such as which players
the user currently has on their teams). This is only sufficient for development,
as other features, notably including email verification, are missing.