Update model.go to be less db-dependent.
Top level model.go is meant to be the business logic representation
of the data structures and shouldn't be a 1-1 mapping with the db
representation.  The backend could change, for example to<->from
postgres and google sheets, or something else. In particular, the
first version of the draft will likely be entirely in memory, with
a bid sent to googlesheets before completely refreshing state.