Create structure for bidding CLI program.
Previous project setup was targeted more at using postgresql to
import the google sheets data and then operate upon it. Given the
low likelihood of migrating all users from google sheets to a new
web app for this year's draft, focus shifted to the auto-bid. To
do this, we added a loop to the CLI which polls google sheets for
updates, identifies if the provided team is the one to make bids,
gets those bids, and then places them.  This is the structure for
the entire bidding process. Other changes actually implementing it
to follow.

Given the new focus on using google sheets, this change also
continued the distinction between database and business logic, and
removed unused postgresql files. These can be added back later when
they actually serve a purpose.