Add STATUS file.
This was in local-only storage for awhile, as a place to keep the
"next thing" for when I neglected the project for months at a time
and lost all context.  Adding it to src control in case it's ever
interesting to look back on what would essentially be a diary of
meta-commentary in addition to actual commits.