Add test fixtures and first test.
Add internal package which also hosts testutils, which can be useful for
adding tests as the business logic proceeds. First test implemented for
the team composition function, which found an error wherein we were not
updating the gap data structure between selected start counts and the max
allowed.  Add that as well (yay testing!).