Add basic bid structure and initial converstion.
Create bidToCell() helper function and associated test which converts a
bid object to the appropriate cell in the googlesheet where it is to be
written.

Meanwhile, also add the skeleton for doing batch bids to reduce API usage
and record the etag value so that we can abort bid placement if the sheet
data has changed since last reading.